+++ src/provisioning/Schema.ts
@@ -13,7 +13,9 @@ const matrixRoomIdSchema: JSONSchemaType<string> = {
 
 const remoteRoomChannelSchema: JSONSchemaType<string> = {
     type: "string",
-    pattern: "^([#+&]|(![A-Z0-9]{5}))[^\\s:,]+$",
+    // According to https://www.rfc-editor.org/rfc/rfc1459#section-1.3
+    // eslint-disable-next-line no-control-regex
+    pattern: "^#([^:\\x00-\\x1F\\s,]){1,199}$",
 };
 
 const remoteRoomServerSchema: JSONSchemaType<string> = {
@@ -27,6 +29,19 @@ const opNickSchema: JSONSchemaType<string> = {
 
 const keySchema: JSONSchemaType<string> = {
     type: "string",
+    // The regex was designed with the following considerations:
+    // - It cannot start with ':' because that would indicate a trailing
+    //   parameter and we treat key exclusively as a middle parameter.
+    // - Commas are disallowed to prevent multiple keys, as we do not support
+    //   joining multiple channels simultaneously.
+    // - Space is disallowed because it signifies the end of the parameter. We
+    //   use \s instead of a literal space to also exclude some Unicode
+    //   whitespace characters out of precaution.
+    //   (see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_expressions/Cheatsheet)
+    // - Control characters (ASCII 00-1F) are excluded to prevent issues with
+    //   printing and parsing reliability. This is more stringent than
+    //   recommended by RFC2812.
+    pattern: "^(?!:)[^\\x00-\\x1F\\s,]*$",
     nullable: true,
 };